Rename `Middleware\Handler` namespace to `Middleware\Unauthorized\Handler`.
