+++ src/Middleware/UnauthorizedHandler/HandlerInterface.php
@@ -12,7 +12,7 @@
  * @since         1.0.0
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
-namespace Authorization\Middleware\Handler;
+namespace Authorization\Middleware\UnauthorizedHandler;
 
 use Authorization\Exception\Exception;
 use Psr\Http\Message\ResponseInterface;
