+++ tests/TestCase/Middleware/UnauthorizedHandler/CakeRedirectHandlerTest.php
@@ -13,19 +13,39 @@
  * @license       https://opensource.org/licenses/mit-license.php MIT License
  */
 
-namespace Authorization\Test\TestCase\Middleware\Handler;
+namespace Authorization\Test\TestCase\Middleware\UnauthorizedHandler;
 
 use Authorization\Exception\Exception;
-use Authorization\Middleware\Handler\RedirectHandler;
+use Authorization\Middleware\UnauthorizedHandler\CakeRedirectHandler;
 use Cake\Http\Response;
 use Cake\Http\ServerRequest;
+use Cake\Routing\Router;
 use Cake\TestSuite\TestCase;
 
-class RedirectHandlerTest extends TestCase
+/**
+ * Description of CakeRedirectHandlerTest
+ *
+ * @author Robert Pustułka <robert.pustulka@gmail.com>
+ */
+class CakeRedirectHandlerTest extends TestCase
 {
-    public function testHandleRedirection()
+    public function setUp()
     {
-        $handler = new RedirectHandler();
+        parent::setUp();
+
+        Router::reload();
+        Router::fullBaseUrl('http://localhost');
+        Router::connect(
+            '/login',
+            ['controller' => 'Users', 'action' => 'login'],
+            ['_name' => 'login']
+        );
+        Router::connect('/:controller/:action');
+    }
+
+    public function testHandleRedirectionDefault()
+    {
+        $handler = new CakeRedirectHandler();
 
         $exception = new Exception();
         $request = new ServerRequest();
@@ -40,9 +60,10 @@ class RedirectHandlerTest extends TestCase
         $this->assertEquals(302, $response->getStatusCode());
         $this->assertEquals('/login?redirect=http%3A%2F%2Flocalhost%2F', $response->getHeaderLine('Location'));
     }
-    public function testHandleRedirectionWithQuery()
+
+    public function testHandleRedirectionNamed()
     {
-        $handler = new RedirectHandler();
+        $handler = new CakeRedirectHandler();
 
         $exception = new Exception();
         $request = new ServerRequest();
@@ -52,7 +73,33 @@ class RedirectHandlerTest extends TestCase
             'exceptions' => [
                 Exception::class,
             ],
-            'url' => '/login?foo=bar'
+            'url' => [
+                '_name' => 'login'
+            ]
+        ]);
+
+        $this->assertEquals(302, $response->getStatusCode());
+        $this->assertEquals('/login?redirect=http%3A%2F%2Flocalhost%2F', $response->getHeaderLine('Location'));
+    }
+
+    public function testHandleRedirectionWithQuery()
+    {
+        $handler = new CakeRedirectHandler();
+
+        $exception = new Exception();
+        $request = new ServerRequest();
+        $response = new Response();
+
+        $response = $handler->handle($exception, $request, $response, [
+            'exceptions' => [
+                Exception::class,
+            ],
+            'url' => [
+                '_name' => 'login',
+                '?' => [
+                    'foo' => 'bar'
+                ],
+            ]
         ]);
 
         $this->assertEquals(302, $response->getStatusCode());
@@ -61,7 +108,7 @@ class RedirectHandlerTest extends TestCase
 
     public function testHandleRedirectionNoQuery()
     {
-        $handler = new RedirectHandler();
+        $handler = new CakeRedirectHandler();
 
         $exception = new Exception();
         $request = new ServerRequest();
@@ -71,23 +118,10 @@ class RedirectHandlerTest extends TestCase
             'exceptions' => [
                 Exception::class,
             ],
-            'url' => '/users/login',
             'queryParam' => null,
         ]);
 
         $this->assertEquals(302, $response->getStatusCode());
-        $this->assertEquals('/users/login', $response->getHeaderLine('Location'));
-    }
-
-    public function testHandleException()
-    {
-        $handler = new RedirectHandler();
-
-        $exception = new Exception();
-        $request = new ServerRequest();
-        $response = new Response();
-
-        $this->expectException(Exception::class);
-        $handler->handle($exception, $request, $response);
+        $this->assertEquals('/login', $response->getHeaderLine('Location'));
     }
 }
